Fix static driver equality check
A recent refactor of the config objects (which simplified the
repetitive __eq__ methods) missed updating the static driver.  This
could lead to infinite recursion (as the static driver explicitly
called super().__eq__ which itself called __eq__).

This updates the driver to use the new framework, and it also adds
a check to a unit test which exercises it.

Change-Id: I0f443cde147e46d73112025cd2f819159a8b4f86