Fix AWS driver equality check
A recent refactor of the config objects (which simplified the
repetitive __eq__ methods) missed updating the AWS driver.  This
could lead to infinite recursion (as the AWS driver explicitly
called super().__eq__ which itself called __eq__).

This updates the driver to use the new framework, and it also adds
a unit test which exercises it.

Change-Id: I3c612e2784de1ffd1642587ba6017e36bebd8d67