Add requestor_data field to node request
Zuul currently keeps track of what buildset+job for which any given
node request was submitted in memory.  To support multiple schedulers,
we need to put that in persistent storage.  We could create a new
kind of object in ZK, but it would be simpler if we could simply store
a memo on the node request itself.

This adds a requestor_data field to the node request object in
ZooKeeper.  It is opaque data to Nodepool, but will allow Zuul to record
the information it needs.

Change-Id: Ie6e29e88f15709829d24dcd063d6afb37778a72e