First ensure ssh connection is valid before scanning keys
We have a network appliance we test via nested virt. While the outer
node is live and the port we nodescan is open, the nested node is still
booting up SSHd.  Which causes nodescan to return:

  paramiko.ssh_exception.SSHException: Error reading SSH protocol banner

until SSHd is properly running.

Perviously we set out boot-timeout to 5 mins, to allow for the nested
SSHd to come online properly. This should restore that functionality.

Change-Id: I7f43530ee77a81f7c969d548190a71bfb9b03455
Signed-off-by: Paul Belanger <pabelanger@redhat.com>