Update patch set 1
Patch Set 1:

Indeed it seems to not work

 https://zuul.opendev.org/t/zuul/build/78112c57351741158c8b96444d59e05a/log/job-output.txt#26416

I think that we expect glean to either write this out (if DNS servers are in metadata) or NetworkManager should write it from DHCP info.  In this case, I would expect the latter.  It seems this isn't happening.

 https://opendev.org/opendev/glean/src/branch/master/glean/cmd.py#L1000

Boot logs are at

 https://zuul.opendev.org/t/zuul/build/78112c57351741158c8b96444d59e05a/log/instances/76674f19-e474-4273-bee5-7175e4601978/console.log#889

 [  141.729532] NetworkManager[709]: <info>  [1627960378.9050] dns-mgr[0x5575c3f05130]: init: dns=default,systemd-resolved rc-manager=symlink

I think that this means networkmanager should try to write /etc/resolv.conf.  But we are using configdrive.  I don't think we're passing in any dns info; dhcp isn't providing it, so i think this is why we are not getting dns info.

Patch-set: 1