Update Docker and bindep for Bullseye base images
Similar to Zuul (I71182e9d3e6e930977a9f983b37743ee3300ec91), the base
images have updated to Bullseye.

This updates various things to get a building Bullseye image.

Also the current master branch mixes dependencies on 3.7 in the
Dockerfile with 3.8 in the zuul jobs -- fix this to be 3.7 and make
sure we depend on the bullseye based images for the jobs too.

Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/806318
Change-Id: I21cfbd3935e48be4b92591ea36c7eed301230753