Temporarily pin back to older python base image
The recent bullseye update requires changes in both nodepool and
dib, which creates a circular depend. In order to deal with that,
pin back to a previous buster-based copy that happens to be laying
around. Then we can land the nodepool and dib updates that are needed.
Then we can revert this.

Change-Id: I30f238081918b209b8ad1cf791b239079c20ab57