Deregister launchers when restarting them
In the NodePool run thread we will restart PoolWorker threads if we
discover they are no longer alive. However, when we do that we don't
call stop() on the object first which deregisters the PoolWorker
launcher from zookeeper. This could lead to leaking launchers in
zookeeper. Fix this by calling stop() on the PoolWorker before starting
a new one.

Change-Id: I07788987b0f0d56fa195d29ba7413b452a30d034