Add user_data field to Node
This will allow users to record information about their use of the
node.  Nodepool will ignore this.

This is similar to the comment field, except that it won't show up
in the node listing output.  It's opaque data to nodepool.

This is also similar to requestor_data for the node request.

We can use this in Zuul to record the tenant/project information,
then just use the existing nodepool node records to calculate
resource usage by tenant.  Without this, we would need to create
a nearly parallel data structure for that purpose alone.  Bonus;
it will let Zuul include autohold node resource usage in its
stats (if desired).

Change-Id: If382ca1c554ddd2ef9be8ad87049b133c8b3873c