Update patch set 2
Patch Set 2:

I have some operational thoughts on this mainly regarding the non-ephemeral locks and local allocation table. Please correct me if my assumptions are wrong.

The non-ephemeral locks require that the my-id of the provider is persisted accross restarts. This is not a problem but on k8s based launchers this needs to be handled explicitly by adding a persistent volume or generate the id on startup. So this would require a change an operator needs to take care of.

Further the optimization using the local allocation table requires that there are never two of the same providers in the system at the same time. We usually do rolling restarts which means that our providers are for a short time duplicated (with different ids however). This would not be possible with this provider. So I think a recreate update strategy needs to be chosen when working with this provider.

Further the recreate strategy means that the provider is unavailable for a short time. Thus to avoid node failures we'd need to run two differently named but equally configured metastatic providers (that don't get restarted at the same time, aka adding a pod disruption budget spanning both deployments).

tldr:
This provider seems to need special care and thoughts at least when run under k8s. This is a change compared to the other providers where throwing them into a k8s deployment using default options just works also when doing frequent rollouts (e.g. do rollouts on config changes).

So I'm wondering if we need to document deployment assumptions/implications somewhere?

Patch-set: 2
CC: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>