kubernetes: Add startup-timeout
Allows the timeout while waiting for a container to be specified. On
constrained clusters, the default timeout of 300 seconds may be too
short to allow builds to succeed since containers may not be able to
start immediately due to resource constraints.

Update the way that the driver waits for a node to become available to
wait as long as the node is in the "Pending" or "Unknown" state, instead
of waiting as long as it is not in the "Running" state. This allows the
driver to detect container errors without having to wait for the entire
timeout, lessening the potential impact of having a longer timeout.

Change-Id: I3d56fc098cf5130d67b4570b55be4520f85a3ba9