Update patch set 2
Patch Set 2:

Tobias Henkel wrote:

> I have some operational thoughts on this mainly regarding the non-
> ephemeral locks and local allocation table. Please correct me if my
> assumptions are wrong.
>
> The non-ephemeral locks require that the my-id of the provider is
> persisted accross restarts. This is not a problem but on k8s based
> launchers this needs to be handled explicitly by adding a persistent
> volume or generate the id on startup. So this would require a change
> an operator needs to take care of.

The my-id is f'NodePool:metastatic:{self.provider.name}' so it will be the same each time regardless.  I don't think we need to handle anything specially there.

> Further the optimization using the local allocation table requires
> that there are never two of the same providers in the system at the
> same time. We usually do rolling restarts which means that our
> providers are for a short time duplicated (with different ids
> however). This would not be possible with this provider. So I think
> a recreate update strategy needs to be chosen when working with this
> provider.

It's true that the in-memory allocation table makes it very unadvisable to run more than one launcher with this provider config.  But I think with Nodepool in general we have the assumption that only one launcher will run a given provider.  Here's the design from the v3 spec:

  The initial implementation will assume only one launcher is running
  for each provider in order to avoid complexities involving quota
  spanning across launchers, rate limits, and how to prevent request
  starvation in the case of multiple launchers for the same provider
  where one is handling a very large request. However, future work may
  enable this with more coordination between launchers in zk.

We may not have added that recommendation to the Nodepool docs, however.  If not, we should.  I think with that assumption in place, this driver doesn't violate it, so from that perspective, we're not adding an additional requirement.

Having said that, you are an expert, and you know what can and does happen if you run more than one provider, and you can certainly choose to make the decision that the benefits outweigh the risks (ie, that you don't mind getting a few quota errors during a restart).  I'm not opposed to highlighting why this particular driver is even riskier than normal to help folks in that position.  :)

Finally, if we want to reduce barriers to running multiple launchers simultaneously, we could move the allocation table into ZK.  But that seems like an enhancement beyond our minimum requirements in nodepool right now.

> Further the recreate strategy means that the provider is unavailable
> for a short time. Thus to avoid node failures we'd need to run two
> differently named but equally configured metastatic providers (that
> don't get restarted at the same time, aka adding a pod disruption
> budget spanning both deployments).

You could run two metastatic providers that supply the same label (and have the same backing provider) but have different provider names.  That matches the typical nodepool way of providing fault-tolerance across clouds.

> tldr: This provider seems to need special care and thoughts at least
> when run under k8s. This is a change compared to the other providers
> where throwing them into a k8s deployment using default options just
> works also when doing frequent rollouts (e.g. do rollouts on config
> changes).

Nodepool is designed to avoid necessitating redeployment on config updates.  It constantly reloads its config files, so you should only need to restart a launcher for a code update.

In zuul-operator, the only time we start/stop/restart an operator is to add or remove a provider (because we run one launcher per provider) and to upgrade the image.

> So I'm wondering if we need to document deployment
> assumptions/implications somewhere?

This is all very useful feedback, and I'm very interested in talking with you more about how you run nodepool and whether we can reduce downtime further.  To make this change actionable, how about the following?

1) Add a note to the docs for the metastatic driver noting the consequences for running more than one launcher with the provider configured.  (Do this in a new patchset for this change.)
2) Add a note to the nodepool docs in general warning about running more than one launcher with a given provider configured.  (Do this in a separate change.)

Patch-set: 2