Rely on bullseye image for podman in container image
Debian bullseye includes podman, so we can drop the kubic repository
(which as I write this is having synchronization errors) in favor of
just using the version from the underlying OS.

The comment mentioned install-recommends being useful for obtaining
some important packages.  They seem to be in the depends line in
debuntu, so it should be sufficient to just install 'podman'.

Change-Id: I95af348a5bf7a05770e8cb47eaffe44b909efbdb