Fix openshift dep for arm64 docker image builds
We tell pip to prefer wheels if present when building arm64 images. We
do this because building wheels for some package installs is extremely
slow on qemu emulated arm64 image builds. Unforatunetly openshift 0.0.1
is the only openshift package version to have a wheel built for it. This
means that we install 0.0.1 which is not compatible with our current
nodepool driver.

This could be a problem for other installations with older openshift
libraries as well as pip won't auto update them. To fix this correctly
we set the lower bound for the openshift library to 0.11.2 which is what
we are using on x86 and is expected to work.

Change-Id: Ida4a27b66fd875166dce2880c7a3c8e0e8aafe94