launcher: ensure we can recover in case of zk outage
If we get a zk outage when we start the node request, we end up
with a locked node request with the requested status. A ConnectionLoss
exception is raised during the run() but is never caught. Because of
that the new request handler is never stored in the request_handlers
list.
The clean up of the node request is never done. The status will
remain "requested" until nodepool is restarted.

Change-Id: I2487d10c3093693b79b3321064ae19c4a242cdd9