Update patch set 10
Patch Set 10:

Andreas,

If you have two metastatic labels (x and y), and they were configured with the same backing-label, then they would use the same node, space permitting.

Of course the point of the driver is for the backing nodes to be on a cloud, and for the cloud to scale.  So assuming that the backing label is configured to support 2 nodes, then if you requent an "x" node, and then a "y" node, they will use the same backing node.  A third concurrent request for x or y would cause a new backing node to be created.

Patch-set: 10