builder: don't use DibImageFile to construct dib image path
This refactors the _buildImage function to not use DibImageFile to
construct the path to the dib output files.

DibImageFile represents one on-disk image.  The dib build argument is
different -- you give it the basename in "-o" and the output types,
and it creates "basename.ext" where "ext" represents the varioius
output types it produced.

This converts _buildImage to the simple thing and makes it a bit
clearer this is a basename.

Change-Id: I214581ad80b7740e7ca749b574672d2c33b92474