Fix tenant-quota for providers without label quota infos
Some providers do not implement the quotaNeededByLabel method, e.g. the
static driver. When having tenant-resource-limits enabled, this method
gets called to find out the resources needed when fulfilling a node
reuquest. In the case of the static driver, this errors out with an
AttributeError.
To circumvent this, this change provides a standard implementation in
the lowest level Provider class which simply returns a zero-value
QuotaInformation object for quotaNeededByLabel() if not overridden.

Change-Id: I76449379d29e87260884c16bfce4c47093eafb2d