Update patch set 2
Patch Set 2: Code-Review+2

Foremost, I think if a user says "this tenant has an instance quota" that ought to be able to work with all drivers.  And if some drivers can also support cores, ram, etc, great.  So I think it's time to at least add in a really basic quota functionality to all drivers.

And test it.  I'm going to push up a change to do that and rebase this on top of it.

If we then also want to keep this change as a fallback so that at least we don't crash if we omit that from some new driver, I'm okay with it.  But I'm also okay with considering that a bug we should fix, in which case maybe we shouldn't land this change and instead fix the quota support in whatever driver is missing it.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Attention: {"person_ident":"Gerrit User 31083 \u003c31083@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}