Stop running siblings jobs and pin SDK
The OpenStackSDK project is in the process of making a breaking
API change and has stopped running this job on master.  Without
an interest in adjusting Nodepool to match before the release,
it is no longer useful to run this job on this repo.  Stop running
it until the sdk project is ready to re-establish co-gating.

We also set a constraint to avoid installing SDK 1.0.0 since we
expect that to include the breaking API change.

Change-Id: If9f45e24a71f349a85e94150e6a4d9ee9672173b