Add additional tests to the aws driver
This adds several new tests:

This exercises the new ipv6 functionality.  The boto3 mocks do not
support assigning an ipv6 address on creation, so we adopt a
unit-test style approach and verify the calling side and our
processing of the results separately.  This uses the patching
scheme that we just removed in the previous commit, but alters it
to simply record the call arguments so we can validate them outside
the patched method.

This adds a diskimage upload test to the AWS driver.  Because moto
doesn't support the create_image method, we need to add a fake for
that.  Boto's practice of creating proxy objects makes that hard to
do, so some helper methods are added on the adapter class to make
it easier for the tests to override.

The responses in the fake are based on a recorded session.

This adds a test of leaked resource cleanup, including the automatic
tagging of untagged resources from image import tasks.

Change-Id: I0626061b246e9c52b08c49394d4d22b46beeff7a