Add provider image config to statemachine image upload
This adds an extra argument to the provider image upload method
so that it can have access to the provider image configuration
which it may need in order to obtain extra information such as
the architecture.

It also adds the upload number to the image name format so that
we may name image uploads by their number like we do instances.

Change-Id: I0f47b4443d86f021641f315af4b69da26c4713a6