Retain created_time in node requests
This attribute is set by Zuul, and it is useful for Zuul to be able
to report statistics.  It also may be useful for Nodepool for the
same purposes, or potentially for future information display.

To facilitate all of the above, retain the attribute across read/
write cycles in Nodepool.

Change-Id: Ic1352d9a25a1f5f3fbbab1313ec832b151ac1cc2