openstack: Keep node request on quota failure
Instead of failing to allocate a node on a quota failure, invalidate the
quota cache and sleep to try again later. This ensures that if OpenStack
is using some sort of dynamic quota mechanism, the "pressure" of
nodepool asking for a node remains as long as nodepool still needs one
so that the quota can be increased if appropriate.

Change-Id: I8d5c0c75cebdbd2ff2dc1ba69fb70152a4cb0fe4