Pass requestor data to Nodes
Propagate the NodeRequest.requestor to a new property Node.requestor
which usually holds the zuul_system_id, or "NodePool:min_ready".
This is set initially when first creating a node, and is updated when a
ready node is assigned to a request. This is to always know for which
requestor is a node got allocated.

Change-Id: Ifd54a94bae39f31a70bdedce384a64c2a04495c1