Fix flavor handling for openstacksdk 1.0
The newest openstacksdk release returns an object of type
Flavor instead of a dict, which does have an id field, but that might
not correspond to an existing flavor, so we cannot find it in our cache.
Check for the presence of the attributes that we really need and as a
last resort skip quota calculation instead of failing.

Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>
Change-Id: I09b03916598ff147d4be210a27a59799c23a2041