Support K8s 1.22.0 and newer.
The rbac.authorization.k8s.io/v1beta1 API was deprecated and replaced
with rbac.authorization.k8s.io/v1. Version 1.22.0 of Kuberenetes removes
the deprecated API which means Nodepool needs to stop using it.
According to the docs [0] version 1.8 and newer support the new APIs.

To do this we need to update the openshift client which updates the
kubernetes client. This appears to be necessary because updating the
version specifier in our requests was still being handled by the beta
version handler in the old kubernetes client library required by the old
openshift library.

Update the explicit versions used and bump client libraries up to their
current version. Add a release note indicating that K8s 1.8 and newer is
now required. Additionally we unpin minikube in testing to ensure we have
test coverage of this change against newer Kuberentes.

[0] https://kubernetes.io/docs/reference/using-api/deprecation-guide/#rbac-resources-v122

Story: 2009974
Change-Id: Ic21725efc8088e7dfb0777c2c96c742182cdbf93