Delete AWS image builds
The AWS driver adapter was not actually deleting the image build
from AWS when the builder requested it (it was simply missing
the implementation of the method).

This went unnoticed because as long as a launcher is running, it
will act to clean up leaked images.  But it's still better to
have the builder do so synchronously.

Add a test which verifies the behavior.

Change-Id: I0bea930847ded45574bf53ae098b9926d4924107