Measure rate limits from start times
The RateLimiter utility used by several drivers measures time from
the exit of one context manager to the beginning of the next.  This
means that if the API call itself takes substantial time, it will
be added to the expected delay.  In other words, if the expected rate
is one API call every two seconds, and the API call itself takes
one second, the actual rate will be one call every three seconds.

To more closely approximate the overall expected rate, measure from
start to start, so that the duration of the API call itself is
included.

Change-Id: Ia62a6bfa6a3e6cac65f0c20179edcfbc94a5dcc5