Fix object refresh in the AWS adapter
Because we cache the instance listing with a TTL, it is possible
for the refresh method (which is intended to return the current
version of the supplied instance object) to encounter a listing
which does not yet include a newly created instance.  In this case
it should just return the supplied argument rather than None which
could cause callers to error out.

Change-Id: I30c04b0f7bdb6f61d2f2fce2037a0959c16dcba9