Fix unknown label in metastatic driver
If a label is removed from the configuration while a node still
exists, the periodic cleanup performed by the metastatic driver
will raise an AttributeError exception when trying to access
the grace_time attribute, since the label pointer is None.

To address this, treat a missing label as if it has a grace time
of zero seconds.

This adds a test which simulates the issue.

This also adds an extra log entry for when a metastatic backing
node slot is deallocated is added so that we log both the allocation
and deallocation.

Change-Id: I0c104a2fe9874e2cd30e2bf2f2227569a73be243