Suppress component registry logging in command
Users of the "nodepool" command don't need to see the component
registry logs at info level (which output at least a line for each
connected component).  Set the minimum level to warning to avoid
that.

The component registry may still be useful for command-line use
in the future, so we leave it in place rather than disabling it
entirely.

Change-Id: I8c0937d7304ddc536773cf74fc40bbf6e79918d4