Avoid collisions after ZK image data import
When image data are imported, if there are holes in the sequence
numbers, ZooKeeper may register a collision after nodepool-builder
builds or uploads a new image.  This is because ZooKeeper stores
a sequence node counter in the parent node, and we lose that
information when exporting/importing.  Newly built images can end
up with the same sequence numbers as imported images.  To avoid this,
re-create missing sequence nodes so that the import state more
closely matches the export state.

Change-Id: I0b96ebecc53dcf47324b8a009af749a3c04e574c