Fix ignore_provider_quota in statemachine drivers
The statemachine driver copied the quota handling from OpenStack,
which also included a check for the presence of the ignore_provider_quota
configuration attribute.  This attribute does not exist in any of
the statemachine drivers, and due to the way the check was constructed,
it meant that we defaulted to ignoring the provider quota rather than
actually checking it.

We should perform this check so that providers can exclude themselves
if there is no possibility of satisfying a request.  This corrects the
check.

Change-Id: I7ced3f749a1646ecbb2b80f93b26e61a4e8cd69a