Add python3.10 testing
This is half expected to fail because the arm64 image build should end
up being incredibly slow on the qemu emulated arm64 setup. However, that
said it appears a number of the historically problematic python
dependencies that do compilations in the nested environment ship aarch64
wheels now. Use this change as a data gathering exercise to see how bad
the docker image buidls are and whether or not we have other
incompatibilities via testing.

Change-Id: I7fb585bc5ccc52803eea107e76dddf5e9fde8646